playground/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "ipykernel>=6.29.5",
    "mermaid-py>=0.7.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "semantic-kernel>=1.27.2",
    "uvicorn>=0.34.0",
]
